refactor(core-packages): fix package hygiene findings
- Remove broken "./back-compat" export from @appstrate/core (pointed to a
  non-existent file; any importer would hard-crash) + drop its stale
  EXEMPT_PATHS entry in lint-afps-1x-keys.ts.
- Drop phantom runtime deps (@appstrate/db, @appstrate/shared-types,
  drizzle-orm) from @appstrate/connect — none are imported, and they
  dragged the db/env credential graph into a credential-isolated package.
- Move runtime Zod schemas/fns (tokenUsageSchema, accumulateTokenUsage,
  modelCostSchema, orgSettingsSchema) out of the types-only @appstrate/
  shared-types into @appstrate/core; update callers. shared-types now
  re-exports types only.
