+++ packages/connect/package.json
@@ -21,9 +21,6 @@
     "@appstrate/afps-runtime": "workspace:*",
     "@appstrate/afps-shared": "workspace:*",
     "@appstrate/core": "workspace:*",
-    "@appstrate/db": "workspace:*",
-    "@appstrate/env": "workspace:*",
-    "@appstrate/shared-types": "workspace:*",
-    "drizzle-orm": "^0.45.2"
+    "@appstrate/env": "workspace:*"
   }
 }
