+++ packages/shared-types/src/index.ts
@@ -261,10 +261,7 @@ export type EnrichedSchedule = ScheduleWireDto & {
 import type { OrgRole } from "@appstrate/core/permissions";
 export type { OrgRole };
 
-export const orgSettingsSchema = z.object({
-  apiVersion: z.string().optional(),
-  dashboardSsoEnabled: z.boolean().optional(),
-});
+import type { orgSettingsSchema } from "@appstrate/core/permissions";
 export type OrgSettings = z.infer<typeof orgSettingsSchema>;
 
 export interface OrganizationMember {
@@ -436,49 +433,9 @@ export interface OrgPackageItemDetail extends OrgPackageItem {
   has_unarchived_changes?: boolean;
 }
 
-// --- Model Cost Types ---
+// --- Token Usage Types ---
 
-/**
- * Zod validator for {@link ModelCost} (the type itself lives in
- * `@appstrate/core/module`). `cacheRead` / `cacheWrite` are optional —
- * providers without prompt caching simply omit them.
- */
-export const modelCostSchema = z.object({
-  input: z.number().nonnegative(),
-  output: z.number().nonnegative(),
-  cacheRead: z.number().nonnegative().optional(),
-  cacheWrite: z.number().nonnegative().optional(),
-});
-
-/**
- * Token usage as reported by an LLM provider for a single completion call.
- * Wire shape consumed by the runner-event ingestion route and any
- * cost-accounting consumer.
- */
-export const tokenUsageSchema = z.object({
-  input_tokens: z.number().nonnegative().optional(),
-  output_tokens: z.number().nonnegative().optional(),
-  cache_creation_input_tokens: z.number().nonnegative().optional(),
-  cache_read_input_tokens: z.number().nonnegative().optional(),
-});
 export type { TokenUsage };
-
-/**
- * In-place accumulator for {@link TokenUsage} totals.
- *
- * Adds every field of `addition` onto `total`. Optional fields default to
- * zero on both sides — `undefined` on `addition` is a no-op, and the
- * cache-creation / cache-read totals are coerced to a numeric zero on
- * `total` so subsequent reads always yield a number.
- */
-export function accumulateTokenUsage(total: TokenUsage, addition: TokenUsage): void {
-  total.input_tokens = (total.input_tokens ?? 0) + (addition.input_tokens ?? 0);
-  total.output_tokens = (total.output_tokens ?? 0) + (addition.output_tokens ?? 0);
-  total.cache_creation_input_tokens =
-    (total.cache_creation_input_tokens ?? 0) + (addition.cache_creation_input_tokens ?? 0);
-  total.cache_read_input_tokens =
-    (total.cache_read_input_tokens ?? 0) + (addition.cache_read_input_tokens ?? 0);
-}
 
 // --- Package Version Types ---
